api_yamdb/requirements.txt
asgiref==3.3.2
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.9.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
pytest-django 
pytest-pythonpath